// src/cubism-common/CubismSpec.d.ts
export namespace CubismSpec {
// ...
  export interface Motion {
    /**
     * File name.
     */
    File: string;
    /**
     * [Optional] Time of the Fade-out for motion easing in seconds.
     */
    FadeOutTime?: number;
    /**
     * [Optional] Time of the Fade-In for motion easing in seconds..
     */
    FadeInTime?: number;
    /**
     * [Optional] Audio files playback with motion.
     */
    Sound?: string;
  }
// ...
}